human-perception-place-pulse/requirements.txt
pandas
Pillow
torch
torchvision
huggingface_hub 
transformers
minio
python-dotenv
sqlalchemy
psycopg2-binary 